Moved database-related code to the database folder.
